Created simple folder struture of project

+++ backend/src/main/java/com/shubham/codeforgeai/repository/CodeFileRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CodeFileRepository extends JpaRepository<CodeFile, Long> {
+public interface CodeFileRepository extends JpaRepository<CodeFile, Integer> {
 
     List<CodeFile> findByProject(Project project);
